feat(UART): add functionality for user to send char input in terminal
User can connect with a serial communication terminal i.e. Putty to send
a single character to interact with the LEDs on the Tiva C board.

Implement the Tx and Rx functions with busy-wait polling method for
simplicity.

Note that setting the UART_CC_R = 0x5 fixes the problem of junk
characters being shown on the terminal. Presumably this is due to baud
rate mismatch while using the system clock of CC_R = 0, where as the
PIOSC (when set to 5) is set to 16 MHz for sure.