feat(UART): add functionality to take user input strings in terminal
Improves the code to now read in strings instead of just characters from
the user. The characters are shown real time so the user is also
visually able to see what they are typing into the terminal and freely
use the backspace key to delete and edit what they are typing.