feat(UART): Add if else branches to turn on the LED
The user types in one of the 8 possible colors from the RGB LED on the
launchpad and is now able to turn it on by typing the desired string in
the serial terminal and not just a character. The string does not have
to be case sensitive to account for user error.

BUG: Weird behavior when sometimes typing in a color turns the LED off