fix(UART): LED bug turning off when transmitting correct color string
Solved a major bug where sometimes the LED would not turn on the
corresponding color and would instead turn it off even though the user
would type in the correct color. It was an issue with the buffer saving
the backspace character in the buffer array so it destroyed the
integrity of the string. Added a check for the '0x7F' ASCII code because
that is the 'DEL' character and how Putty reads in backspaces.

Add delay after the uart_0_initialization to let the clock settle.
This solved the junk characters from appearing without it.