feat(ring_buffer): Improve logic to use ring buffer
Set up a ring buffer and test function to make sure it works properly.
This is necessary for the next step that will be to implement UART
interrupts. Currently it is a work in progress.

The logic in the ring buffer functions has been sorted out. Test
function has not been finished.

TODO: Finish implementing the ring buffer test function and confirm it
works properly before moving onto the UART interrupts.