feat(led): Implement colors.c to turn on corresponding LED color
Three functions were implemented to turn on one of the 8 possible LED
colors on the launchpad.

get_color()
It calls the uart0_interrupt_get_string() function to build the string
in the colors[] array and returns an enum color to be passed into the
next function.

led_turn_on_color()
It uses the enum color parameter as a switch statement argument to turn
on the corresponding color on the launchpad LED.

strings_compare_colors_case_insensitive()
This function was brought over from the uart_busy_wait.c file. It was
slightly modified to return a boolean to improve code readability.

TODO: Add string prompts on the terminal such as "Enter color: " or
"invalid color entered..."