feat(prompt_menu): Add terminal prompts to ask user to input a color
Implemented two functions in colors.c that calls on
uart0_interrupt_send_char() to print prompts onto the terminal screen.
It prompts the user to enter a valid color, and the user is able to see
the characters they type in live.

Once the user enters the string, the terminal will print "received" then
echo the correct color the user typed in, or output "Invalid!"

Bug: Edge case, if the input string is too long such as "greeeeeeeeen",
then the "Received:" doesn't clear out the remaining characters and
instead prints out "Received:eeeeeen"

Tried the following solutions
1) "\rReceived:\n" to clear out the line prior to printing Received,
   doesn't work
2) "Received:             \n" adding spaces to clear out the characters
   but then it prints out random characters in lieu of that.