bug(prompt_menu): Fix menu prompts by showing user input in full
The program shows the user input and doesn't clear the line so it's
clear what color they chose and why they got an invalid received if they
made a typo or typed in a color that didn't exist.

There was weird behavior with the code when the ring buffer wasn't big
enough to be able to handle the amount of characters that needed to be
printed onto the menu prompt. Tested each module extensively and
confirmed the ring buffer is working as it should as well. It seems to
be a buffer overflow issue.